Rename E2E tests because they're not really E2E
License: MIT
